Move (errors.h)concat --> (utils.h)qflex::utils::concat.
